refactor: re-use SDK functionality where possible
