--- contracts/utils/events.ts
@@ -1,24 +1,0 @@
-import {ContractTransaction} from 'ethers';
-import {Interface, LogDescription} from 'ethers/lib/utils';
-
-export async function findEvent<T>(tx: ContractTransaction, eventName: string) {
-  const receipt = await tx.wait();
-
-  const event = (receipt.events || []).find(event => event.event === eventName);
-
-  return event as T | undefined;
-}
-
-export async function findEventTopicLog<T>(
-  tx: ContractTransaction,
-  iface: Interface,
-  eventName: string
-): Promise<LogDescription & (T | LogDescription)> {
-  const receipt = await tx.wait();
-  const topic = iface.getEventTopic(eventName);
-  const log = receipt.logs.find(x => x.topics[0] === topic);
-  if (!log) {
-    throw new Error(`No logs found for the topic of event "${eventName}".`);
-  }
-  return iface.parseLog(log) as LogDescription & (T | LogDescription);
-}
